+++ features/support/env.rb
@@ -31,6 +31,16 @@ Before('@skip_agp4_1_or_higher') do |scenario|
   skip_this_scenario() if is_above_or_equal_to_target(410)
 end
 
+Before('@skip_agp3_5') do |scenario|
+  skip_this_scenario() if equals_target(350)
+end
+
+def equals_target(target)
+  version = ENV["AGP_VERSION"].slice(0, 5)
+  version = version.gsub(".", "")
+  return version.to_i == target
+end
+
 def is_above_or_equal_to_target(target)
   version = ENV["AGP_VERSION"].slice(0, 5)
   version = version.gsub(".", "")