+++ src/test/kotlin/com/bugsnag/android/gradle/UploadRequestClientTest.kt
@@ -1,5 +1,6 @@
 package com.bugsnag.android.gradle
 
+import com.bugsnag.android.gradle.internal.LegacyUploadRequestClient
 import org.junit.Assert.assertEquals
 import org.junit.Test
 
@@ -15,40 +16,40 @@ class UploadRequestClientTest {
 
     @Test
     fun testRequestDiffVersionInfo() {
-        val client = UploadRequestClient()
+        val client = LegacyUploadRequestClient()
         var requestCount = 0
         val request = {
             requestCount += 1
             ""
         }
-        client.makeRequestIfNeeded(info, "{}", request)
-        client.makeRequestIfNeeded(info.copy(apiKey = "40fadb0123094f"), "{}", request)
+        client.makeRequestIfNeeded(info, "{}".hashCode(), request)
+        client.makeRequestIfNeeded(info.copy(apiKey = "40fadb0123094f"), "{}".hashCode(), request)
         assertEquals(2, requestCount)
     }
 
     @Test
     fun testRequestDiffPayload() {
-        val client = UploadRequestClient()
+        val client = LegacyUploadRequestClient()
         var requestCount = 0
         val request = {
             requestCount += 1
             ""
         }
-        client.makeRequestIfNeeded(info, "{}", request)
-        client.makeRequestIfNeeded(info, "[]", request)
+        client.makeRequestIfNeeded(info, "{}".hashCode(), request)
+        client.makeRequestIfNeeded(info, "[]".hashCode(), request)
         assertEquals(2, requestCount)
     }
 
     @Test
     fun testRequestSameInfo() {
-        val client = UploadRequestClient()
+        val client = LegacyUploadRequestClient()
         var requestCount = 0
         val request = {
             requestCount += 1
             ""
         }
-        client.makeRequestIfNeeded(info, "{}", request)
-        client.makeRequestIfNeeded(info, "{}", request)
+        client.makeRequestIfNeeded(info, "{}".hashCode(), request)
+        client.makeRequestIfNeeded(info, "{}".hashCode(), request)
         assertEquals(1, requestCount)
     }
 }