+++ CHANGELOG.md
@@ -5,11 +5,18 @@ improvements compared to previous versions of the plugin, such as:
 
 - Supporting [up-to-date checks](https://docs.gradle.org/current/userguide/more_about_tasks.html#sec:up_to_date_checks)
 - Supporting [task configuration avoidance](https://docs.gradle.org/current/userguide/task_configuration_avoidance.html)
+- Supporting the [configuration cache](https://docs.gradle.org/current/userguide/configuration_cache.html)
 - Avoidance of unnecessary work
 
 See the [Upgrade Guide](./UPGRADING.md) for migration instructions.
 
 ### Changes
+
+Support configuration caching
+[#257](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/257)
+
+Change how bugsnag plugin is disabled for build variants
+[#255](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/255)
 
 Convert plugin extension to use property syntax
 [#251](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/251)
@@ -17,11 +24,10 @@ Convert plugin extension to use property syntax
 Deduplicate unnecessary upload requests for APK splits
 [#248](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/248)
 
-Add autoUpdateBuildUuid flag to prevent manifest UUID generation
-[#249](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/249)
-
 Switch to OkHttp for networking
 [#247](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/247)
+[#268](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/268)
+[#269](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/266)
 
 Specify shared object files as input for ndk task
 [#243](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/243)